pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kvseed-guard"
version = "0.1.0"
description = "Head-agnostic KV seeding and logit gating middleware for LLM runtimes"
readme = "README.md"
authors = [{name = "kvseed-guard contributors"}]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.1.0",
    "typer>=0.9",
    "rich>=13.6",
    "pynacl>=1.5",
    "cryptography>=41",
    "PyYAML>=6.0",
    "pydantic>=2.0",
    "numpy>=1.24",
    "uvloop>=0.17; platform_system != 'Windows'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "mypy>=1.8",
    "ruff>=0.2.0",
    "build>=1.0",
    "pre-commit>=3.5",
]

[project.scripts]
kvguard = "kvseed_guard.cli.main:app"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
config_file = "setup.cfg"

[tool.ruff]
config = "setup.cfg"
